Change EnemyPhaseStartCondition to use expressions; fix some fields in Enemy and DeathAction using floats instead of ints

+++ BulletHellMaker/DeathAction.h
@@ -170,8 +170,8 @@ public:
 	sf::Color getColor();
 	Animatable getAnimatable();
 	bool getLoopAnimatable();
-	float getMinParticles();
-	float getMaxParticles();
+	int getMinParticles();
+	int getMaxParticles();
 	float getMinDistance();
 	float getMaxDistance();
 	float getMinLifespan();
